Rename cli/cli -> cli/index
